feat: use slot
