+++ src/app/(public)/(home)/@profile/(.)profile/[slug]/page.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/app/(public)/profile/[slug]/page";
