+++ src/app/(public)/(home)/@profile/page.tsx
@@ -1,0 +1,9 @@
+import { cn } from "@/libs/utils";
+
+export default function Page() {
+  return (
+    <div className={cn("px-3", "py-2", "text-5xl", "font-bold")}>
+      This space is intended for publication use.
+    </div>
+  );
+}
